Adding Authentication and CSRF protection

+++ controllers/admin.js
@@ -32,7 +32,7 @@ exports.postAddProduct = (req, res, next) => {
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find()
+  Product.find({ userId: req.user._id })
     .then(products => {
       res.render('admin/products', {
         prods: products,
@@ -54,15 +54,17 @@ exports.postEditProduct = (req, res, next) => {
 
   Product.findById(prodId)
     .then(product => {
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect('/');
+      }
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
       product.imageUrl = updatedImageUrl;
-      return product.save();
-    })
-    .then(() => {
-      console.log('Updated Product');
-      res.redirect('/admin/products');
+      return product.save().then(result => {
+        console.log('Updated Product');
+        res.redirect('/admin/products');
+      });
     })
     .catch(err => console.log(err));
 };
@@ -91,7 +93,7 @@ exports.getEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findByIdAndRemove(prodId)
+  Product.findByIdAndRemove({ _id: prodId, userId: req.user._id })
     .then(() => {
       console.log('PRODUCT DELETED');
       res.redirect('/admin/products');
